Move Enterprise-only kube-controllers settings to an Enterprise configuration
The base kube-controllers configuration only carries fields the base render reads.
